% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{readAKNcsv}
\alias{readAKNcsv}
\title{Read in AKN files into data frame}
\usage{
readAKNcsv(csvfiles)
}
\arguments{
\item{csvfiles}{List of csv files to convert to data frame.}
}
\value{
A data frame combining the loaded csv files.
}
\description{
Read in AKN files into data frame
}

