% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{readrefuge}
\alias{readrefuge}
\title{Read refuge shapefile into sp data object}
\usage{
readrefuge(refugedsn, layer)
}
\arguments{
\item{refugedsn}{File location of refuge shapefile (.shp)}

\item{layer}{Name of layer within shapefile}
}
\value{
sp polygon for refuge
}
\description{
Read refuge shapefile into sp data object
}

