% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{recentAKNs}
\alias{recentAKNs}
\title{Return most recent AKNs as a data table}
\usage{
recentAKNs(AKNs)
}
\arguments{
\item{AKNs}{Data frame giving AKN locations}
}
\value{
Data frame with most recent AKNs
}
\description{
Return most recent AKNs as a data table
}

