% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{refugeAKNfraction}
\alias{refugeAKNfraction}
\title{Return proportion of AKNs that are within the refuge}
\usage{
refugeAKNfraction(refuge, AKNs)
}
\arguments{
\item{refuge}{Refuge sp data object}

\item{AKNs}{Data frame giving AKN locations}
}
\value{
Scalar proportion of AKNs within refuge
}
\description{
Return proportion of AKNs that are within the refuge
}

