% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{refugeAKNs}
\alias{refugeAKNs}
\title{Return sp points of AKNs within the refuge}
\usage{
refugeAKNs(refuge, AKNs)
}
\arguments{
\item{refuge}{Refuge sp data object}

\item{AKNs}{Data frame giving AKN locations}
}
\value{
Data frame of observation points within refuge
}
\description{
Return sp points of AKNs within the refuge
}

