% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{writeAKNcsv}
\alias{writeAKNcsv}
\title{Write out AKNs as a csv file}
\usage{
writeAKNcsv(AKNs, fname)
}
\arguments{
\item{AKNs}{Data frame giving AKN locations}

\item{fname}{Output filename}
}
\value{
(empty)
}
\description{
Write out AKNs as a csv file
}

