% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_functions.R
\name{writeAKNshp}
\alias{writeAKNshp}
\title{Write out AKNs as shapefile}
\usage{
writeAKNshp(AKNs, layer, fname)
}
\arguments{
\item{AKNs}{Data frame giving AKN locations}

\item{layer}{Output layer name}

\item{fname}{output filename}
}
\value{
(empty)
}
\description{
Write out AKNs as shapefile
}

